Move finality/grandpa/warp_sync to sync/warp_sync/verifier
